[REG-1059] use new config methods for co-op bot
